refactor(core): move one port's batch-slot orchestration into tyutool-core
`run_batch_auth_slot` was the largest function left in src-tauri's batch.rs at
270 lines, and the spec called moving it a rewrite: the orchestration state
machine had to be split from event emission first. That split already existed.
The signature has taken `emit: &dyn Fn(serde_json::Value)` rather than an
`AppHandle` since it was written, with a comment saying why — so the function
had no Tauri reference at all, and the only thing tying it to this crate was
`logs::BatchAuthTraceWriter`, which P4b-1 moved.

So this is a move. `AllocatorSession`, `SlotSessionGuard`, `BatchAuthStartConfig`
and the slot body become `tyutool_core::batch_slot`, behind the `excel` feature
because the signature names `ExcelRowAllocator`. Renamed to `run_batch_slot` to
keep it distinct from core's existing `authorize::run_batch_auth_slot`, which is
the single device transaction one layer below it. Both tests move with it.

`BatchAuthStartConfig`'s fields become `pub`: the struct is deserialized from
the frontend and read by the command layer, which is now across a crate
boundary.

batch.rs drops from 927 to 543 lines. What remains is the Tauri command layer —
`AppHandle`, `State`, the per-port thread pool, emit forwarding — which by the
crate-boundary table mostly belongs there. Whether any of it should move is a
design question, not a size question: core now offers a full single-port run, so
a CLI would only need its own concurrency shell. AGENTS.md and the spec both say
to answer that before moving anything further.
